Move App file to src folder
