+++ src/App.tsx
@@ -1,0 +1,50 @@
+import React from 'react';
+import {
+  SafeAreaView,
+  ScrollView,
+  StatusBar,
+  StyleSheet,
+  Text,
+  TextInput,
+  View,
+} from 'react-native';
+
+const App = () => {
+  return (
+    <SafeAreaView>
+      <StatusBar />
+      <ScrollView contentInsetAdjustmentBehavior="automatic">
+        <View style={styles.sectionContainer}>
+          <Text style={styles.sectionTitle}>Text Input</Text>
+          <Text style={styles.sectionDescription}>
+            All the text goes into that box below
+          </Text>
+          <TextInput style={styles.textInput} />
+        </View>
+      </ScrollView>
+    </SafeAreaView>
+  );
+};
+
+const styles = StyleSheet.create({
+  sectionContainer: {
+    marginTop: 32,
+    paddingHorizontal: 24,
+  },
+  sectionTitle: {
+    fontSize: 24,
+    fontWeight: '600',
+  },
+  sectionDescription: {
+    marginTop: 8,
+    fontSize: 18,
+    fontWeight: '400',
+  },
+  textInput: {
+    marginTop: 16,
+    borderColor: '#ccc',
+    borderWidth: 1,
+  },
+});
+
+export default App;
